fix: collapse duplicate Advanced/MarkerMenu stories into one file
Two story files declared title 'Advanced/MarkerMenu', so Storybook derived the
same story ID for both and the static build aborted, failing publish-docs.

Port the two play interaction tests (DisallowedMarker, CharacterMarkerRemoveRow)
from src/stories/advanced/ into the co-located file, restyled to its conventions,
then delete the duplicate. The files were not redundant -- only the deleted one
carried play assertions.
